Convert search transforms to the Search plugin
Moves BM25SearchTransform/RegexSearchTransform/BaseSearchTransform from
`src/fastmcp/server/transforms/search/` to
`src/fastmcp/server/plugins/search/` and adds a `Search` plugin (with
`SearchConfig`) that exposes BM25 or regex search as a single
`plugins=[...]` entry. The transforms stay importable from the old
path via a deprecation shim.

This is the first Phase 1 conversion under FMCP-20 — the template
every other first-party feature (prompts-as-tools, resources-as-tools,
openapi, skills, code-mode) will follow.

Usage:

    from fastmcp import FastMCP
    from fastmcp.server.plugins.search import Search, SearchConfig

    mcp = FastMCP("s", plugins=[Search(SearchConfig(strategy="regex"))])
    # or with a dict for JSON/YAML loading:
    mcp = FastMCP("s", plugins=[Search({"always_visible": ["help"]})])

Existing `from fastmcp.server.transforms.search import ...` paths
still work; the package emits a DeprecationWarning pointing at the new
location.
